Fixes & improves custom XML file loading example.
The example:
GridPanel > Miscellaneous > XML File Loading Custom

Had long inliner handlers and was also broken on Ext.NET 3, so that the
expanded rows' data was not displayed in the inner panels.

That was due to the gridpanel being called its load() method after
expanding the row.

To fix that, proper load() handling was written and raw data bound to a
placeholder into the inner grid's store.

The long handlers were turned into JavaScript functions and bound to
shorter 'Fn'. This also allowed to add descriptive comments to the
JavaScript code.

This addesses issue extnet/Ext.NET#1252 item 51.