Improves BorderLayout's Collapsed Region Image example.
The panels were being expanded as floating quick-expand objects when
clicked by the title. This didn't disable the images as the example
stated, so could be confuse for users learning Ext.NET.

Added TitleCollapse="false" and Floatable="false" to the panels inside
the window on the example.

Affected example is
Layout > BorderLayout > Collapsed Region Image

This addresses issue extnet/Ext.NET#1252 item 64.