Adds a workaround for the drop rules being ignored.
Daniil has contributed with the fix after he found it happening on the
code. As the fix from Ext.NET DLL will only be available after next
public release, this workaround was applied based on his proposed
solution.

Related github issues: #11.
This was a side effect of extnet/Ext.NET#1330.