Reviews Project file support for Visual Studio
Removed old rules that allowed it to work on very old versions of Visual
Studio and that deterministicly broken the project to any new versions of
Visual Studio. Notably Owin startup class that was limited only to VS
versions explicitly specified.

This addresses issue #15 item 408 (Stock ticker example not working).