Fixes the slider overview example.
Sets up the change event function to stop execution if the last
component of the page is not yet ready, preventing it from making
references to undefined components.

This triggers now because the 'change' event is called also once the
component is loaded, before the next ones in the page could have been
defined.

Related github issues:
- extnet/Ext.NET#1516
- extnet/Ext.NET.Examples#21 item 198.