Reviews the FilterHeader custom filter for mapped field example.
Adds support for filtering other columns, still with a regexp custom
behavior, that will match the text anywhere in the columns instead of
the beginning, as default by the FilterHeader component.

It will match by the column name when deciding which filter to use.

Also cleaned up the code from debugging 'console.log()' calls.