Renames all references to Theme.Default and 'blue' to 'Classic'.
The theme has not been the default Ext.NET theme for a long long time,
and in Ext.NET 5 the theme name was changed to reflect its actual theme
name in Ext JS, which is 'classic'.

References to 'blue' as the name as well as the enum reference,
Ext.Net.Theme.Default, changed to Ext.Net.Theme.Classic.

This comes as a consequence of extnet/Ext.NET#1602.